Move Bar an Session events to application layer
