+++ core/src/main/java/com/tungstun/core/application/bar/event/BarDeleted.java
@@ -1,0 +1,4 @@
+package com.tungstun.core.application.bar.event;
+
+public record BarDeleted(Long id) {
+}
