+++ core/src/main/java/com/tungstun/core/application/session/event/SessionCreated.java
@@ -1,0 +1,4 @@
+package com.tungstun.core.application.session.event;
+
+public record SessionCreated(Long id, Long barId) {
+}
