+++ core/src/main/java/com/tungstun/core/application/session/event/SessionDeleted.java
@@ -1,0 +1,4 @@
+package com.tungstun.core.application.session.event;
+
+public record SessionDeleted(Long id, Long barId) {
+}
